Added functions
Can access multiple user with their perspective data

+++ MODEL/PersonalInfo.cs
@@ -7,12 +7,12 @@ namespace MODEL
         public String Gender { get; set; }
         public DateTime DateofBirth { get; set; }
         public String placeofBirth { get; set; }
-        public int mobileNo { get; set; }
+        public long mobileNo { get; set; }
         public String emailAddress { get; set; }
         public String residentialAddress { get; set; }
         public String permanentAdress { get; set; }
 
-        public PersonalInfo(string fullName,string gender, DateTime dateofBirth, string placeofBirth, int mobileNo, string emailAddress, string residentialAddress, string permanentAdress) : base(fullName)
+        public PersonalInfo(string fullName,string gender, DateTime dateofBirth, string placeofBirth, long mobileNo, string emailAddress, string residentialAddress, string permanentAdress) : base(fullName)
         {
             this.fullName = fullName;
             Gender = gender;
